Add cancel attribute and styles to Button component
Introduced a `cancel` property to Button components for representing cancelation actions, along with styling that visually differentiates these buttons with a red theme. This enhancement is motivated by the need to alert users to potentially destructive actions or to otherwise signify a negative operation, increasing the UI's intuitiveness. It should be noted that utilizing this new feature requires passing a `cancel` prop to the Button component.

Refs #4567